agregando detalles antes de cagarla

+++ js/main.js
@@ -24,7 +24,6 @@ function verificacion(){
   contU=0;
   userStorage.forEach(usuario => {
     if( usuarioIngresado === usuario.nombre && contraseniaIngresada === usuario.contraseña){
-      debugger
       localStorage.setItem(0,JSON.stringify(usuario));
       setTimeout(detectarPagina(),2000);
       window.location="index.html";
@@ -32,7 +31,13 @@ function verificacion(){
     else{
       contU++;
       if(contU===userStorage.length){
-        alert("ingreso algun dato mal");
+        Swal.fire({
+          position: 'top-end',
+          icon: 'error',
+          title: 'Ingreso un dato erroneo',
+          showConfirmButton: false,
+          timer: 1500
+        })
       }
     }
   })
@@ -40,30 +45,35 @@ function verificacion(){
 }
 
 function registrar (){
+
   usuarioIngresado = document.getElementById("usuario").value;
   contraseniaIngresada = document.getElementById("pass").value;
-  persona = new Usuarios (usuarioIngresado,contraseniaIngresada);
-  respuestaAdmin = confirm("Desea ser admin?")
-  if(respuestaAdmin){
-    persona.admin = 1;
-  }
-  else{
-    persona.admin = 0;
-  }
-  
-  if(localStorage.getItem(3)===null){
-  userStorage.push(persona);
-  userStorageJson = JSON.stringify(userStorage);
-  localStorage.setItem(3,userStorageJson);   
-  }
-  else{
-    userStorage = JSON.parse(localStorage.getItem(3));
-    userStorage.push(persona);
-    userStorageJson = JSON.stringify(userStorage);
-    localStorage.setItem(3,userStorageJson);   
-  }
+  comparador = detectarUsuariosYaRegistrados(usuarioIngresado); // uso la funcion para saber si ya el usuario ya existe
+  if (comparador===true){       
 
+    persona = new Usuarios (usuarioIngresado,contraseniaIngresada);
+    respuestaAdmin = confirm("Desea ser admin?")
+    if(respuestaAdmin){
+      persona.admin = 1;
+    }
+    else{
+      persona.admin = 0;
+    }
+    
+    if(localStorage.getItem(3)===null){
+      userStorage.push(persona);
+      userStorageJson = JSON.stringify(userStorage);
+      localStorage.setItem(3,userStorageJson);   
+    }
+    else{
+      userStorage = JSON.parse(localStorage.getItem(3));
+      userStorage.push(persona);
+      userStorageJson = JSON.stringify(userStorage);
+      localStorage.setItem(3,userStorageJson);   
+    }
+    
   }
+}
   
   function detectarAdmin (){
     parametroEnJson = localStorage.getItem("0");
@@ -72,11 +82,11 @@ function registrar (){
     if (parametro.admin === 1){
       dive.innerHTML = ` 
       <form action="agregarObjeto">
-      <h3>Agregar Articulos</h3>
-      <p>Nombre</p><input type="text" id="nombre-nuevo-producto">
-      <p>precio</p><input type="number" id="precio-nuevo-producto">
-      <p>stock</p><input type="number" id="stock-nuevo-producto">
-      <button type="button" onclick="registrar_Producto()" id="boton_registrar_producto">Añadir</button>
+      <h3 class = "agregarObjetosTitulo">Agregar Articulos</h3>
+      <input class="inputProductos" placeholder="Nombre" type="text" id="nombre-nuevo-producto">
+      <input class="inputProductos" placeholder="Precio" type="number" id="precio-nuevo-producto">
+      <input class="inputProductos" placeholder="Stock" type="number" id="stock-nuevo-producto">
+      <button class="botonAgregarObjeto" type="button" onclick="registrar_Producto()" id="boton_registrar_producto">Añadir</button>
       </form> <hr>
       `;
     }
@@ -138,7 +148,7 @@ function registrar (){
           liProd.setAttribute("id","row_"+elemento.id);
           liProd.innerHTML = `<div"><h3 class="nombre">${elemento.nombre}</h3></div>
           <div><p class="precio" >precio: ${elemento.precio}</p></div>
-          <div><p class="stock">stock: ${elemento.stock}</p></div>
+          <div ><p class="stock">stock: ${elemento.stock}</p></div>
           <div class="id"><p>id: ${elemento.id}</p></div>
           <button type="button" id="addCarrito" onclick="addCarrito(${elemento.id})">añadir al Carrito</button>`
           ulDeProductos.appendChild(liProd);
@@ -179,6 +189,30 @@ function registrar (){
       );
       mandarCarrito(producto);   
     } 
+    function restCarrito(id){        //escanea los datos del producto seleccionado y los convierte en un objeto
+      carrito = recuperarCarrito();
+      elementoCarrito = carrito.filter(element => element.id===id);
+      objetoAcomparar = elementoCarrito[0];
+      if (objetoAcomparar.cantidad===1){
+        columna=document.getElementById("colum_"+id);
+        columna.innerHTML="";
+        posicion= carrito.map(elemento => elemento.id).indexOf(id);
+        carrito.splice(posicion,1);
+      }
+      else{
+        objetoAcomparar.cantidad--;
+
+      }
+        carritoJson = JSON.stringify(carrito);
+        localStorage.setItem("1",carritoJson);
+        productosArray = JSON.parse(sessionStorage.getItem(0));
+        ObjetoComprado = productosArray[id];
+        ObjetoComprado.stock++;
+        sessionStorage.setItem(0,JSON.stringify(productosArray));
+        actualizarCarrito();
+        detectarElementosEnCarrito();
+        mostrarProductos();
+      } 
     function mandarCarrito(producto){ //detecta si el producto ya esta en el carrito. en el caso afirmativo aumenta la variable "cantidad" en 1 del objeto. en el caso negativo agrega el objeto al array "carrito" y cambia la variable "cantidad" a 1
 
       if (carrito.some(element => element.id === producto.id)){
@@ -189,12 +223,27 @@ function registrar (){
         let cantidad=1;
         producto.cantidad=cantidad;
         carrito.push(producto);
-        alert("se agrego producto con exito");
+        const Toast = Swal.mixin({  //Sweetalert
+          toast: true,
+          position: 'top-end',
+          showConfirmButton: false,
+          timer: 1500,
+          timerProgressBar: true,
+          didOpen: (toast) => {
+            toast.addEventListener('mouseenter', Swal.stopTimer)
+            toast.addEventListener('mouseleave', Swal.resumeTimer)
+          }
+        })
+        
+        Toast.fire({
+          icon: 'success',
+          title: 'Se añadio al carrito'
+        })
       
       }
     ;
-      carritoJson = JSON.stringify(carrito);
-      localStorage.setItem("1",carritoJson);
+        carritoJson = JSON.stringify(carrito);
+        localStorage.setItem("1",carritoJson);
       actualizarCarrito();
       detectarElementosEnCarrito();
 
@@ -209,9 +258,11 @@ function registrar (){
       carrito.forEach(producto => {
         precioFinal = producto.precio*producto.cantidad;
         elemento = document.createElement("tr");
+        elemento.setAttribute("id","colum_"+producto.id)
         elemento.innerHTML = `<th scope="row">${producto.nombre}</th>
         <td>${producto.precio}</td>
         <td>${producto.cantidad}</td>
+        <td> <button type="button" onclick="addCarrito(${producto.id})" >+</button>  <button type="button" onclick="restCarrito(${producto.id})" >-</button></td>
         <td id="precio_carrito">${precioFinal}</td>`
         divCarrito.appendChild(elemento);
       });
@@ -237,9 +288,10 @@ function registrar (){
       precios.forEach(element => {
         c=element.textContent*1+c;
       })
-      elemento.innerHTML = `<th scope="row" colspan="2"></th>
+      elemento.innerHTML = `<th scope="row" colspan="3"></th>
       <td><h4>Total:</h4></td>      
-      <td><h4>${c}</h4></td>  `;
+      <td><h4>${c}</h4></td>
+      `;
       body.appendChild(elemento);
     }
     async function detectarPagina(){   //detecta si estamos posicionado en el index para recien ahi ejecutar las funciones 
@@ -273,7 +325,11 @@ function registrar (){
     }
     function iniciar(){//agregue esta funcion porque la funcion "detectar pagina" no funciona si la ip del ordenador cambia, solo funciona con el live server
       if(document.title==='Carrito'){
-
+        Swal.fire(
+          'Bienvenido',
+          'Es algo divertido',
+          'success'
+        )
         detectarAdmin();
         mostrarProductos();
         recuperarCarrito();
@@ -291,3 +347,23 @@ function registrar (){
       return array;
     }
 
+    function detectarUsuariosYaRegistrados (nombre){        //sirve para analizar si el usuario ingresado ya existe
+      
+      if (localStorage.getItem(3)===null){
+        return true;
+      }
+      else{
+
+        userStorage = JSON.parse(localStorage.getItem(3));
+        if (userStorage.filter(elemento => elemento.nombre === nombre).length>0){
+          alert("El nombre de usario ya esta registrado")
+          return false
+        }
+        else{
+          return true;
+        }
+      }
+    }
+    function realizar_compra(){
+      window.location="Final.html";
+    }
